+++ src/screen/dashboard.rs
@@ -1882,7 +1882,35 @@ impl Dashboard {
                 }
 
                 let (pane_to_split, pane_to_split_state) = {
-                    if self.focus.window == self.main_window()
+                    if matches!(
+                        config.pane.split_axis,
+                        config::pane::SplitAxis::LargestShorter
+                    ) && let Some((pane, pane_state)) = panes
+                        .iter()
+                        .filter_map(|(window, id, state)| {
+                            (window == self.main_window())
+                                .then_some((id, state))
+                        })
+                        .reduce(
+                            |(acc_pane, acc_pane_state), (pane, pane_state)| {
+                                let pane_area = pane_state.size.width
+                                    * pane_state.size.height;
+                                let acc_pane_area = acc_pane_state.size.width
+                                    * acc_pane_state.size.height;
+
+                                if pane_area > acc_pane_area
+                                    || (pane_area == acc_pane_area
+                                        && pane == self.focus.pane)
+                                {
+                                    (pane, pane_state)
+                                } else {
+                                    (acc_pane, acc_pane_state)
+                                }
+                            },
+                        )
+                    {
+                        (pane, pane_state)
+                    } else if self.focus.window == self.main_window()
                         && let Some(pane_state) =
                             self.panes.main.panes.get(&self.focus.pane)
                     {
@@ -1904,7 +1932,8 @@ impl Dashboard {
                     config::pane::SplitAxis::Vertical => {
                         pane_grid::Axis::Vertical
                     }
-                    config::pane::SplitAxis::Shorter => {
+                    config::pane::SplitAxis::Shorter
+                    | config::pane::SplitAxis::LargestShorter => {
                         if pane_to_split_state.size.height
                             < pane_to_split_state.size.width
                         {