+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Added:
 - Ability to hide and theme `kick` server messages
 - Setting to control whether or not to restore on launch the panes that were open when Halloy last closed
 - Setting for split axis chosen as the shorter dimension of the focused pane (new default behavior)
+- Setting for split axis chosen as the shorter dimension of the largest splittable pane
 
 Fixed: